int id 2

+++ src/main/java/com/backend/codigobackend/PessoaRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.repository.Repository;
 
 import java.util.List;
 
-public interface PessoaRepository extends Repository<Pessoa, Long> {
+public interface PessoaRepository extends Repository<Pessoa, Integer> {
     List<Pessoa> findAll();
-    Pessoa findById(long id);
+    Pessoa findById(int id);
     Pessoa save(Pessoa pessoa);
     Pessoa findByNome(String nome);
     boolean findByEmail(String email);
